refactor(#1396): split nexus_fast lib.rs (3,587 LOC) into 8 focused modules
Split the monolithic lib.rs into domain-specific modules for better
maintainability and developer ergonomics. The lib.rs hub is now 54 LOC
containing only module declarations and PyO3 re-exports.

Modules created:
- rebac.rs (~1,100 LOC): ReBAC permission engine with string interning
- cache.rs (516 LOC): BloomFilter + L1MetadataCache classes
- grep.rs (497 LOC): SIMD-accelerated content search
- similarity.rs (350 LOC): SIMD cosine/dot/euclidean similarity
- tiger_cache.rs (168 LOC): Roaring Bitmap permission filtering
- glob_filter.rs (134 LOC): Glob pattern matching
- io.rs (124 LOC): Memory-mapped file I/O
- blake3_hash.rs (53 LOC): BLAKE3 content-addressable hashing

DRY improvements in rebac.rs:
- Extracted parse_namespace_configs() replacing 4 duplicated blocks
- Extracted parse_tuples_from_py() replacing 3 duplicated blocks

All 25 function exports + 2 class exports verified.
270/270 sync permission unit tests pass.
No performance regression (all ops sub-millisecond).
