+++ rust/nexus_fast/src/similarity.rs
@@ -1,0 +1,350 @@
+// =============================================================================
+// SIMD-Accelerated Vector Similarity (Issue #952)
+// =============================================================================
+
+use pyo3::prelude::*;
+use rayon::prelude::*;
+use simsimd::SpatialSimilarity;
+
+/// Threshold for parallelization in similarity operations
+const SIMILARITY_PARALLEL_THRESHOLD: usize = 100;
+
+/// Compute cosine similarity between two f32 vectors using SIMD.
+///
+/// Uses SimSIMD for 100x speedup over naive implementation.
+/// ~10ns per 1536-dim vector comparison vs ~1μs naive.
+///
+/// Args:
+///     a: First vector
+///     b: Second vector
+///
+/// Returns:
+///     Cosine similarity (1.0 = identical, 0.0 = orthogonal, -1.0 = opposite)
+#[pyfunction]
+pub fn cosine_similarity_f32(a: Vec<f32>, b: Vec<f32>) -> PyResult<f64> {
+    if a.len() != b.len() {
+        return Err(pyo3::exceptions::PyValueError::new_err(format!(
+            "Vector length mismatch: {} vs {}",
+            a.len(),
+            b.len()
+        )));
+    }
+
+    // SimSIMD returns cosine distance (1 - similarity), so we convert
+    // Use explicit trait syntax to avoid conflict with std::f32::cos
+    <f32 as SpatialSimilarity>::cos(&a, &b)
+        .map(|dist| 1.0 - dist)
+        .ok_or_else(|| pyo3::exceptions::PyValueError::new_err("SIMD cosine computation failed"))
+}
+
+/// Compute dot product between two f32 vectors using SIMD.
+///
+/// Uses SimSIMD for 100x speedup over naive implementation.
+///
+/// Args:
+///     a: First vector
+///     b: Second vector
+///
+/// Returns:
+///     Dot product value
+#[pyfunction]
+pub fn dot_product_f32(a: Vec<f32>, b: Vec<f32>) -> PyResult<f64> {
+    if a.len() != b.len() {
+        return Err(pyo3::exceptions::PyValueError::new_err(format!(
+            "Vector length mismatch: {} vs {}",
+            a.len(),
+            b.len()
+        )));
+    }
+
+    <f32 as SpatialSimilarity>::dot(&a, &b).ok_or_else(|| {
+        pyo3::exceptions::PyValueError::new_err("SIMD dot product computation failed")
+    })
+}
+
+/// Compute squared Euclidean distance between two f32 vectors using SIMD.
+///
+/// Uses SimSIMD for 100x speedup over naive implementation.
+///
+/// Args:
+///     a: First vector
+///     b: Second vector
+///
+/// Returns:
+///     Squared Euclidean distance (L2²)
+#[pyfunction]
+pub fn euclidean_sq_f32(a: Vec<f32>, b: Vec<f32>) -> PyResult<f64> {
+    if a.len() != b.len() {
+        return Err(pyo3::exceptions::PyValueError::new_err(format!(
+            "Vector length mismatch: {} vs {}",
+            a.len(),
+            b.len()
+        )));
+    }
+
+    <f32 as SpatialSimilarity>::l2sq(&a, &b)
+        .ok_or_else(|| pyo3::exceptions::PyValueError::new_err("SIMD L2 computation failed"))
+}
+
+/// Batch cosine similarity: compute similarity of query vs all vectors.
+///
+/// Uses SimSIMD + Rayon for parallel SIMD computation.
+/// Expected 100x speedup: 10ms for 10K vectors → 100μs.
+///
+/// Args:
+///     query: Query vector (f32)
+///     vectors: List of vectors to compare against (f32)
+///
+/// Returns:
+///     List of cosine similarities (same order as input vectors)
+#[pyfunction]
+pub fn batch_cosine_similarity_f32(
+    query: Vec<f32>,
+    vectors: Vec<Vec<f32>>,
+) -> PyResult<Vec<f64>> {
+    if vectors.is_empty() {
+        return Ok(vec![]);
+    }
+
+    // Validate dimensions
+    let query_dim = query.len();
+    for (i, v) in vectors.iter().enumerate() {
+        if v.len() != query_dim {
+            return Err(pyo3::exceptions::PyValueError::new_err(format!(
+                "Vector {} dimension mismatch: expected {}, got {}",
+                i,
+                query_dim,
+                v.len()
+            )));
+        }
+    }
+
+    // Use parallel iteration for large batches
+    let similarities: Vec<f64> = if vectors.len() > SIMILARITY_PARALLEL_THRESHOLD {
+        vectors
+            .par_iter()
+            .map(|v| {
+                <f32 as SpatialSimilarity>::cos(&query, v)
+                    .map(|dist| 1.0 - dist)
+                    .unwrap_or(0.0)
+            })
+            .collect()
+    } else {
+        vectors
+            .iter()
+            .map(|v| {
+                <f32 as SpatialSimilarity>::cos(&query, v)
+                    .map(|dist| 1.0 - dist)
+                    .unwrap_or(0.0)
+            })
+            .collect()
+    };
+
+    Ok(similarities)
+}
+
+/// Top-K similarity search using SIMD.
+///
+/// Finds the K most similar vectors to the query.
+/// Uses parallel SIMD scoring + efficient top-K selection.
+///
+/// Args:
+///     query: Query vector (f32)
+///     vectors: List of vectors to search (f32)
+///     k: Number of top results to return
+///
+/// Returns:
+///     List of (index, similarity) tuples, sorted by similarity descending
+#[pyfunction]
+pub fn top_k_similar_f32(
+    query: Vec<f32>,
+    vectors: Vec<Vec<f32>>,
+    k: usize,
+) -> PyResult<Vec<(usize, f64)>> {
+    if vectors.is_empty() || k == 0 {
+        return Ok(vec![]);
+    }
+
+    // Validate dimensions
+    let query_dim = query.len();
+    for (i, v) in vectors.iter().enumerate() {
+        if v.len() != query_dim {
+            return Err(pyo3::exceptions::PyValueError::new_err(format!(
+                "Vector {} dimension mismatch: expected {}, got {}",
+                i,
+                query_dim,
+                v.len()
+            )));
+        }
+    }
+
+    // Compute all similarities in parallel
+    let mut scores: Vec<(usize, f64)> = if vectors.len() > SIMILARITY_PARALLEL_THRESHOLD {
+        vectors
+            .par_iter()
+            .enumerate()
+            .map(|(i, v)| {
+                let sim = <f32 as SpatialSimilarity>::cos(&query, v)
+                    .map(|dist| 1.0 - dist)
+                    .unwrap_or(0.0);
+                (i, sim)
+            })
+            .collect()
+    } else {
+        vectors
+            .iter()
+            .enumerate()
+            .map(|(i, v)| {
+                let sim = <f32 as SpatialSimilarity>::cos(&query, v)
+                    .map(|dist| 1.0 - dist)
+                    .unwrap_or(0.0);
+                (i, sim)
+            })
+            .collect()
+    };
+
+    // Sort by similarity descending
+    scores.sort_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
+
+    // Truncate to top-K
+    scores.truncate(k);
+
+    Ok(scores)
+}
+
+/// Cosine similarity for int8 quantized vectors using SIMD.
+///
+/// 166x faster than naive + 4x smaller memory footprint.
+/// Use for quantized embeddings to reduce memory and increase throughput.
+///
+/// Args:
+///     a: First vector (i8)
+///     b: Second vector (i8)
+///
+/// Returns:
+///     Cosine similarity
+#[pyfunction]
+pub fn cosine_similarity_i8(a: Vec<i8>, b: Vec<i8>) -> PyResult<f64> {
+    if a.len() != b.len() {
+        return Err(pyo3::exceptions::PyValueError::new_err(format!(
+            "Vector length mismatch: {} vs {}",
+            a.len(),
+            b.len()
+        )));
+    }
+
+    // SimSIMD returns cosine distance, convert to similarity
+    <i8 as SpatialSimilarity>::cos(&a, &b)
+        .map(|dist| 1.0 - dist)
+        .ok_or_else(|| pyo3::exceptions::PyValueError::new_err("SIMD i8 cosine computation failed"))
+}
+
+/// Batch cosine similarity for int8 quantized vectors.
+///
+/// Args:
+///     query: Query vector (i8)
+///     vectors: List of vectors to compare against (i8)
+///
+/// Returns:
+///     List of cosine similarities
+#[pyfunction]
+pub fn batch_cosine_similarity_i8(query: Vec<i8>, vectors: Vec<Vec<i8>>) -> PyResult<Vec<f64>> {
+    if vectors.is_empty() {
+        return Ok(vec![]);
+    }
+
+    let query_dim = query.len();
+    for (i, v) in vectors.iter().enumerate() {
+        if v.len() != query_dim {
+            return Err(pyo3::exceptions::PyValueError::new_err(format!(
+                "Vector {} dimension mismatch: expected {}, got {}",
+                i,
+                query_dim,
+                v.len()
+            )));
+        }
+    }
+
+    let similarities: Vec<f64> = if vectors.len() > SIMILARITY_PARALLEL_THRESHOLD {
+        vectors
+            .par_iter()
+            .map(|v| {
+                <i8 as SpatialSimilarity>::cos(&query, v)
+                    .map(|dist| 1.0 - dist)
+                    .unwrap_or(0.0)
+            })
+            .collect()
+    } else {
+        vectors
+            .iter()
+            .map(|v| {
+                <i8 as SpatialSimilarity>::cos(&query, v)
+                    .map(|dist| 1.0 - dist)
+                    .unwrap_or(0.0)
+            })
+            .collect()
+    };
+
+    Ok(similarities)
+}
+
+/// Top-K similarity search for int8 quantized vectors.
+///
+/// Args:
+///     query: Query vector (i8)
+///     vectors: List of vectors to search (i8)
+///     k: Number of top results to return
+///
+/// Returns:
+///     List of (index, similarity) tuples, sorted by similarity descending
+#[pyfunction]
+pub fn top_k_similar_i8(
+    query: Vec<i8>,
+    vectors: Vec<Vec<i8>>,
+    k: usize,
+) -> PyResult<Vec<(usize, f64)>> {
+    if vectors.is_empty() || k == 0 {
+        return Ok(vec![]);
+    }
+
+    let query_dim = query.len();
+    for (i, v) in vectors.iter().enumerate() {
+        if v.len() != query_dim {
+            return Err(pyo3::exceptions::PyValueError::new_err(format!(
+                "Vector {} dimension mismatch: expected {}, got {}",
+                i,
+                query_dim,
+                v.len()
+            )));
+        }
+    }
+
+    let mut scores: Vec<(usize, f64)> = if vectors.len() > SIMILARITY_PARALLEL_THRESHOLD {
+        vectors
+            .par_iter()
+            .enumerate()
+            .map(|(i, v)| {
+                let sim = <i8 as SpatialSimilarity>::cos(&query, v)
+                    .map(|dist| 1.0 - dist)
+                    .unwrap_or(0.0);
+                (i, sim)
+            })
+            .collect()
+    } else {
+        vectors
+            .iter()
+            .enumerate()
+            .map(|(i, v)| {
+                let sim = <i8 as SpatialSimilarity>::cos(&query, v)
+                    .map(|dist| 1.0 - dist)
+                    .unwrap_or(0.0);
+                (i, sim)
+            })
+            .collect()
+    };
+
+    scores.sort_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
+    scores.truncate(k);
+
+    Ok(scores)
+}
